+++ src/lightning/pytorch/trainer/connectors/data_connector.py
@@ -11,7 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import multiprocessing
 import os
 from dataclasses import dataclass, field
 from typing import Any, Iterable, Optional, Tuple, Union
@@ -25,6 +24,7 @@ from lightning.fabric.utilities.data import (
     _replace_dunder_methods,
     _set_sampler_epoch,
     has_iterable_dataset,
+    suggested_max_num_workers,
 )
 from lightning.fabric.utilities.distributed import DistributedSamplerWrapper
 from lightning.pytorch.overrides.distributed import UnrepeatedDistributedSamplerWrapper
@@ -420,18 +420,16 @@ def _check_dataloader_iterable(
         )
 
 
-def _worker_check(dataloader: object, name: str) -> None:
+def _worker_check(trainer: "pl.Trainer", dataloader: object, name: str) -> None:
     if not isinstance(dataloader, DataLoader):
         return
 
-    num_cpus = multiprocessing.cpu_count()
-    if dataloader.num_workers <= 2 < num_cpus:
+    upper_bound = suggested_max_num_workers(trainer.num_devices)
+    if dataloader.num_workers <= 2 < upper_bound or dataloader.num_workers < 2 <= upper_bound:
         # if changed, update the `filterwarnings` snippet in 'speed.html#num-workers'
         rank_zero_warn(
-            f"The dataloader, {name}, does not have many workers which may be a bottleneck."
-            " Consider increasing the value of the `num_workers` argument`"
-            f" (try {num_cpus} which is the number of cpus on this machine)"
-            " in the `DataLoader` init to improve performance.",
+            f"The '{name}' does not have many workers which may be a bottleneck. Consider increasing the value of the"
+            f" `num_workers` argument` to `num_workers={upper_bound}` in the `DataLoader` to improve performance.",
             category=PossibleUserWarning,
         )
 
@@ -488,7 +486,7 @@ def _process_dataloader(
     dataloader = strategy.process_dataloader(dataloader)
 
     # check the workers
-    _worker_check(dataloader, f"{stage.dataloader_prefix}_dataloader")
+    _worker_check(trainer=trainer, dataloader=dataloader, name=f"{stage.dataloader_prefix}_dataloader")
 
     # add worker_init_fn for correct seeding in worker processes
     _auto_add_worker_init_fn(dataloader, trainer.global_rank)